Add flask test app and confingure test harness
* add test env vars to pytest.ini
* fix tests to use env vars instead
* add initial test with flask test app
