+++ src/gobstuf/__main__.py
@@ -2,9 +2,9 @@ from gobstuf.app import run as run_app
 
 
 def init():
-    if __name__ == "__main__":
-        # Run the app locally
-        run_app()
+    # Run the app locally
+    run_app()
 
 
-init()
+if __name__ == "__main__":
+    init()
