+++ src/tests/pytest.ini
@@ -1,0 +1,11 @@
+[pytest]
+env =
+    ROUTE_PATH_310=/310
+    ROUTE_PATH_204=/204
+    ROUTE_SCHEME=https
+    ROUTE_NETLOC=gob-stuf-tests
+    KEYCLOAK_AUTH_URL=gob-tests
+    KEYCLOAK_CLIENT_ID=fake-keycloak-id
+    GOB_STUF_PORT=8000
+    PKCS12_FILENAME="FAKE_PKCS12_FILENAME"
+    PKCS12_PASSWORD="FAKE_PKCS12_PASSWORD"
