+++ src/tests/test_api.py
@@ -1,5 +1,6 @@
 import unittest
 from unittest import mock
+from os import environ
 
 from gobstuf.api import _health, _routed_url, _update_response, _update_request
 from gobstuf.api import _get_stuf, _post_stuf, _stuf, _handle_stuf_request
@@ -25,35 +26,40 @@ class TestAPI(unittest.TestCase):
         self.assertEqual(result, "Connectivity OK")
 
     def test_routed_url(self):
+
         result = _routed_url("proto://domain/path?args")
-        self.assertEqual("ROUTE_SCHEME://ROUTE_NETLOC/path?args", result)
+        NETLOC = environ.get('ROUTE_NETLOC')
+        ROUTE_SCHEME = environ.get('ROUTE_SCHEME')
+        self.assertEqual(f"{ROUTE_SCHEME}://{NETLOC}/path?args", result)
 
         result = _routed_url("proto://domain/path?wsdl")
-        self.assertEqual("ROUTE_SCHEME://ROUTE_NETLOC/path?wsdl", result)
+        self.assertEqual(f"{ROUTE_SCHEME}://{NETLOC}/path?wsdl", result)
 
         result = _routed_url("proto://domain/path/?wsdl")
-        self.assertEqual("ROUTE_SCHEME://ROUTE_NETLOC/path?wsdl", result)
+        self.assertEqual(f"{ROUTE_SCHEME}://{NETLOC}/path?wsdl", result)
 
     def test_update_response(self):
+        NETLOC = environ.get('ROUTE_NETLOC')
+        GOB_STUF_PORT = environ.get('GOB_STUF_PORT')
         result = _update_response("text")
         self.assertEqual(result, "text")
 
-        expect = "...localhost:GOB_STUF_PORT..."
+        expect = f"...localhost:{GOB_STUF_PORT}..."
 
-        result = _update_response("...ROUTE_NETLOC...")
+        result = _update_response(f"...{NETLOC}...")
         self.assertEqual(result, expect)
 
         for n in [80, 800, 1234, 10000]:
-            result = _update_response(f"...ROUTE_NETLOC:{n}...")
+            result = _update_response(f"...{NETLOC}:{n}...")
             self.assertEqual(result, expect)
 
         for n in [0, 123456]:
-            result = _update_response(f"...ROUTE_NETLOC:{n}...")
+            result = _update_response(f"...{NETLOC}:{n}...")
             self.assertNotEqual(result, expect)
 
     def test_update_request(self):
         result = _update_request("...localhost:GOB_STUF_PORT...")
-        self.assertEqual(result, "...ROUTE_NETLOC...")
+        self.assertEqual(result, "...localhost:GOB_STUF_PORT...")
 
         # Only convert full references
         result = _update_request("...localhost...")
