+++ src/tests/test_certrequest.py
@@ -1,4 +1,5 @@
 import unittest
+from os import environ
 from unittest import mock
 
 from gobstuf.certrequest import cert_get, cert_post
@@ -10,7 +11,7 @@ class MockResponse:
         self.reason = reason
 
 class TestConfig(unittest.TestCase):
-
+    """Test if requests get PKCS12 config from environment"""
     def setUp(self) -> None:
         pass
 
@@ -19,7 +20,11 @@ class TestConfig(unittest.TestCase):
         mock_get.return_value = MockResponse()
 
         response = cert_get("any url")
-        mock_get.assert_called_with("any url", pkcs12_filename="PKCS12_FILENAME", pkcs12_password="PKCS12_PASSWORD")
+        mock_get.assert_called_with(
+            "any url",
+            pkcs12_filename=environ.get("PKCS12_FILENAME"),
+            pkcs12_password=environ.get("PKCS12_PASSWORD")
+        )
 
         self.assertIsInstance(response, MockResponse)
 
@@ -28,10 +33,22 @@ class TestConfig(unittest.TestCase):
         mock_post.return_value = MockResponse()
 
         cert_post("any url", data="any data", headers={"a": 0})
-        mock_post.assert_called_with("any url", data="any data", headers={"a": 0}, pkcs12_filename="PKCS12_FILENAME", pkcs12_password="PKCS12_PASSWORD")
+        mock_post.assert_called_with(
+            "any url",
+            data="any data",
+            headers={"a": 0},
+            pkcs12_filename=environ.get("PKCS12_FILENAME"),
+            pkcs12_password=environ.get("PKCS12_PASSWORD"),
+        )
 
         # headers is a default argument
         response = cert_post("any url", data="any data", headers={})
-        mock_post.assert_called_with("any url", data="any data", headers={}, pkcs12_filename="PKCS12_FILENAME", pkcs12_password="PKCS12_PASSWORD")
+        mock_post.assert_called_with(
+            "any url",
+            data="any data",
+            headers={},
+            pkcs12_filename=environ.get("PKCS12_FILENAME"),
+            pkcs12_password=environ.get("PKCS12_PASSWORD"),
+        )
 
         self.assertIsInstance(response, MockResponse)
